requirements.txt
llama-cpp-python
requests
